v1

+++ posts/models.py
@@ -4,7 +4,7 @@ from django.db import models
 class Post(models.Model):
     position = models.CharField(max_length=200)
     text = models.TextField()
-    price = models.DecimalField(max_digits=10, decimal_places=2)
+    price = models.IntegerField()
     image = models.ImageField(upload_to='posts/', blank=True, null=True)
 
     def __str__(self):
